refactor: extract direct visible thought cleanup
